% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{CellCirclePlot}
\alias{CellCirclePlot}
\title{Generate a circle plot of significant cell-cell connections}
\usage{
CellCirclePlot(
  input.file,
  adj_pval_thresh = 0.05,
  col.set = NULL,
  arrow.size = 0.6,
  arrow.width = 2,
  edge.multi = 0.02
)
}
\arguments{
\item{input.file}{input file}

\item{adj_pval_thresh}{adjusted p-value threshold for defining significance}

\item{col.set}{named vector of colour codes to fill node colours}

\item{lab.weight.thresh}{minimum weight for labelling populations in the plot.}
}
\value{
a ggplot2 object.
}
\description{
Given a file generated by the EvaluateConnections function,
subsets according to adjusted P-value and plots the significant
cell-cell connections. Uses the igraph package for network plotting.
}
\examples{
\dontrun{
CellCirclePlot(input.file = input.file)
}

}
