% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{EvaluateConnections}
\alias{EvaluateConnections}
\title{Perform permutation testing on network edges}
\usage{
EvaluateConnections(file.label, num.permutations = 1e+05,
  num.cores = 1)
}
\arguments{
\item{file.label}{a list of genes}

\item{num.permutations}{number of permutations to perform}

\item{num.cores}{number of cores to use in parallelisation}
}
\value{
NULL - results written to file
}
\description{
Uses permutation testing to identify cell-cell connections that have
path weights greater than would be expected by change.
}
