% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{GenerateEdgeWeights}
\alias{GenerateEdgeWeights}
\title{Generates files of cluster:ligand:receptor:cluster edge weights}
\usage{
GenerateEdgeWeights(
  seurat.object,
  file.label,
  species,
  populations.use = NULL,
  string.dir = NULL,
  string.ver = NULL,
  string.receptors = NULL,
  string.ligands = NULL,
  string.input.map = NULL,
  verbose = FALSE
)
}
\arguments{
\item{seurat.object}{a list of genes}

\item{file.label}{a Seurat object with cluster identities}

\item{species}{expression threshold for considering a gene expressed in a cell}

\item{populations.use}{threshold of percentage of cells expressing the gene in a cluster for it to be considered expressed}

\item{string.dir}{directory for storing STRING data (defaults to working directory)}

\item{string.ver}{version of STRING data-base to use. Not set by default.}

\item{string.receptors}{a list of receptor names that STRING recognises (if cannot map defaults)}

\item{string.ligands}{a list of ligand names that STRING recegnises (if cannot map defaults)}

\item{string.input.map}{named vector of STRING-compatable gene symbols with names corresponding to genes in dataset}

\item{verbose}{whether to print additional information about run (default: FALSE)}
}
\value{
NULL - results written to file
}
\description{
For a Seurat object, calculates
}
