% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_internal.R
\name{get_gene_pair_expression_values}
\alias{get_gene_pair_expression_values}
\title{Identify expression of ligands/receptors among clusters}
\usage{
get_gene_pair_expression_values(
  ligand.receptor.pairs,
  mapping.table,
  cluster.names,
  gene.de.results
)
}
\arguments{
\item{ligand.receptor.pairs}{table of ligand-receptor pairs}

\item{mapping.table}{human:mouse gene ID mappings}

\item{cluster.names}{set of cluster IDs to use}

\item{gene.de.results}{table of expression values produced by calculate_cluster_specific_expression}
}
\value{
a data-frame of results
}
\description{
For a set of ligand-receptor pairs, evaluates whether one of the gene pairs is expressed
for each of the clusters and builds a table mapping each ligand-expressing cluster to
each corresponding receptor-expressing cluster.
}
