% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_internal.R
\name{get_weighted_paths}
\alias{get_weighted_paths}
\title{Determine weights paths between query source and target populations}
\usage{
get_weighted_paths(
  edge.weight.table,
  source.population,
  target.population,
  min.weight = 1.5,
  print.num.connections = FALSE
)
}
\arguments{
\item{edge.weight.table}{a table of edges and weights for building the source:ligand:receptor:target graph}

\item{source.population}{source (ligand-expressing) population}

\item{target.population}{target (receptor-expressing) population}

\item{min.weight}{minimum weight for retaining a source:target path (deafult is 1.5)}

\item{print.num.connections}{whether to print the number of paths passing the min weight threshold (False by default)}
}
\value{
a vector of source:target path weights
}
\description{
Given a table of weights for source:ligands, ligands:receptors and receptors:targets and a query
source and target population, build a weighted, directed graph (using the igraph package) and pull
out summed paths between the query source and target populations. Retain paths above a user-provided
minimum weight.
}
\examples{
getWeightedPaths(edge.weight.table, source.population, target.population)
}
