% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{InboundOutboundPlot}
\alias{InboundOutboundPlot}
\title{Plot inbound vs outbound weights}
\usage{
InboundOutboundPlot(sig.paths, this.population, col.set = NULL,
  p.labels = NULL, lab.weight.thresh = 1000)
}
\arguments{
\item{this.population}{a cluster ID contained in the path table.}

\item{col.set}{optional paramater to set colours of the populations.}

\item{lab.weight.thresh}{minimum weight for labelling populations in the plot.}

\item{path.table}{a table of source:ligand:receptor:target paths.}
}
\value{
a ggplot2 object.
}
\description{
Generates a dot plot showing inbound vs outbound weights for all populations.
}
\examples{
PlotTopLigands(path.table = example.table, this.population = '1')
}
