% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_internal.R
\name{make_STRING_table}
\alias{make_STRING_table}
\title{Ligand:receptor association scores}
\usage{
make_STRING_table(ligands, receptors, species = "mouse",
  dir.path = NULL, string.ver = "10")
}
\arguments{
\item{ligands}{a vector of ligands}

\item{receptors}{a vector of receptors}

\item{species}{species to use - either mouse (default) or human}

\item{dir.path}{output directory for storing STRINGdb data. If not provided uses current working directory}
}
\value{
a data-frame containing ligands, receptors and STRING association scores between them.
}
\description{
Generates a table of ligand:receptor association scores from the STRING database
}
\examples{
PlotTopLigands(path.table = example.table, this.population = '1')
}
