% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{NetworkTreePlot}
\alias{NetworkTreePlot}
\title{Generate a network plot linking source:ligand:receptor:targets in tree format}
\usage{
NetworkTreePlot(
  path.file,
  edge.score.file,
  source.population,
  target.populations,
  source.marker.genes = NULL,
  target.marker.genes = NULL,
  ligand.set = NULL,
  receptor.set = NULL,
  population.cols = NULL,
  ligand.col = "#c7d8e8",
  receptor.col = "#d2bcff"
)
}
\arguments{
\item{path.file}{input file}

\item{edge.score.file}{adjusted p-value threshold for defining significance}

\item{source.population}{named vector of colour codes to fill node colours}

\item{target.populations}{minimum weight for labelling populations in the plot.}

\item{ligand.set}{set of ligands to plot}

\item{receptor.set}{set of receptors to plot}

\item{population.cols}{option colour set for cell populations. provide in order of ligand, receptors}

\item{ligand.col}{colour for the row of ligands}

\item{receptor.col}{colour for the row of receptors}
}
\value{
a ggplot2 object.
}
\description{
Given a file generated by the EvaluateConnections function,
subsets according to adjusted P-value and plots the significant
cell-cell connections. Uses the igraph package for network plotting.
}
\examples{
\dontrun{
NetworkTreePlot(path.file = example.file)
}

}
