% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PlotTopLigands}
\alias{PlotTopLigands}
\title{Plot a bar graph of the top ligands}
\usage{
PlotTopLigands(path.table, this.population, col.use = NULL,
  min.weight = 2, max.plot.num = 15)
}
\arguments{
\item{path.table}{a table of source:ligand:receptor:target paths.}

\item{this.population}{a cluster ID contained in the path table.}

\item{col.use}{optional paramater to set colour of the bar graph.}

\item{min.weight}{minimum weight for including paths in the top ligand calculations.}
}
\value{
a ggplot2 object.
}
\description{
Plot a bar graph of the top ligands
}
\examples{
PlotTopLigands(path.table = example.table, this.population = '1')

}
