% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_internal.R
\name{randomise_FC_weights}
\alias{randomise_FC_weights}
\title{Randomise source:ligand and receptor:target weights}
\usage{
randomise_FC_weights(ppi.weights, cluster.ligand.table, receptor.cluster.table)
}
\arguments{
\item{ppi.weights}{a vector of ligands}

\item{cluster.ligand.table}{a vector of receptors}

\item{receptor.cluster.table}{species to use - either mouse (default) or human}
}
\value{
a vector of source:target path weights
}
\description{
Generates a randomisation of the network by shuffling the source:ligand and receptor:target weights,
which are dervied from fold-change expression. The ligand:receptor weights derived from STRINGdb
remain unshuffled. The weights are then summed to represent randomised source:target path weights.
}
\examples{
RandomiseFCWeights(ppi.weights, cluster.ligand.table, receptor.cluster.table)
}
