% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{rescale_node_coordinates}
\alias{rescale_node_coordinates}
\title{Rescale nodes in a ggraph tree for even distribution}
\usage{
rescale_node_coordinates(
  graph,
  scale.by = "Ligand",
  do.update = c("Source", "Ligand", "Receptor", "Target")
)
}
\arguments{
\item{graph}{a ggraph object containing a tree graph.}

\item{scale.by}{a reference row to scale to. Ligand by default.}

\item{do.update}{select graph rows to updates. All by default.}
}
\value{
a ggraph object with node X-coordinates rescaled.
}
\description{
Takes as input a ggraph object containing a tree graph with 1 or more source nodes at the root
connecting to targets in the leaf nodes with connecting ligands and receptors as intermediates.
Rescales the X-coordinates of the nodes to ensure an even spacing. Rescales according to a reference
row - ligands by default, but probably works best when specified as the row with largets number
of nodes. Can select rows to rescale, though all are done by default.
}
\examples{
\dontrun{
rescale_target_node_coordinates(tree.graph)
}
}
