% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{EvaluateConnections}
\alias{EvaluateConnections}
\title{Perform permutation testing on network edges}
\usage{
EvaluateConnections(file.label, num.permutations = 1e+05,
  return.results = FALSE, ncores = 1)
}
\arguments{
\item{file.label}{label for input and output files}

\item{num.permutations}{number of permutations to perform}

\item{return.results}{whether to return the results table (default: FALSE)}

\item{ncores}{number of cores to use in parallelisation}
}
\value{
NULL - results written to file
}
\description{
Uses permutation testing to identify cell-cell connections that have
path weights greater than would be expected by change.
}
