% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{GenerateEdgeWeights}
\alias{GenerateEdgeWeights}
\title{Generates files of cluster:ligand:receptor:cluster edge weights}
\usage{
GenerateEdgeWeights(seurat.object, out.label, species,
  populations.use = NULL)
}
\arguments{
\item{seurat.object}{a list of genes}

\item{out.label}{a Seurat object with cluster identities}

\item{species}{expression threshold for considering a gene expressed in a cell}

\item{populations.use}{Threshold of percentage of cells expressing the gene in a cluster
for it to be considered expressed}
}
\value{
NULL - results written to file
}
\description{
For a Seurat object
}
