% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{GenerateNetworkPaths}
\alias{GenerateNetworkPaths}
\title{Generates files of weighted source:ligand:receptor:target paths}
\usage{
GenerateNetworkPaths(file.label = out.lab, min.weight = 1.5)
}
\arguments{
\item{file.label}{a list of genes}

\item{min.weight}{minimum weight for retaining a path}
}
\value{
NULL - results written to file
}
\description{
Uses files generated with the GenerateEdgeWeights function to generate
weighted paths connecting source(clusters):ligands:receptors:target(clusters).
Paths are weighted by summing the three edges - source:ligand, ligand:receptor and
receptor:target. Paths passing a minimum weight threshold are retained. For
following permuation testing, a 'background' set of paths without thresholding
are also written out.
}
