% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{InboundOutboundPlot}
\alias{InboundOutboundPlot}
\title{Plot inbound vs outbound weights}
\usage{
InboundOutboundPlot(
  input.file,
  col.set = NULL,
  p.labels = NULL,
  p.adj.thresh = 0.05,
  lab.weight.thresh = 1000,
  return.plot = TRUE
)
}
\arguments{
\item{input.file}{a table of source:ligand:receptor:target paths.}

\item{col.set}{optional paramater to set colours of the populations.}

\item{p.labels}{cell-type labels (defaults to labels in input.file)}

\item{p.adj.thresh}{minimum adjusted P-value for including a cell-cell connection}

\item{lab.weight.thresh}{minimum weight for labelling populations in the plot.}

\item{return.plot}{whether to return the ggplot2 object (default: TRUE)}
}
\value{
a ggplot2 object.
}
\description{
Generates a dot plot showing inbound vs outbound weights for all populations.
}
\examples{
\dontrun{
InboundOutboundPlot(path.table = example.table)
}

}
