% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_internal.R
\name{make_STRING_table}
\alias{make_STRING_table}
\title{Ligand:receptor association scores}
\usage{
make_STRING_table(
  ligands,
  receptors,
  species,
  dir.path = NULL,
  string.ver = NULL,
  verbose = FALSE,
  string.receptors = NULL,
  string.ligands = NULL,
  string.input.map = NULL
)
}
\arguments{
\item{ligands}{a vector of ligands}

\item{receptors}{a vector of receptors}

\item{species}{species to use - either mouse (default) or human}

\item{dir.path}{output directory for storing STRINGdb data. If not provided uses current working directory}

\item{string.ver}{STRING version to use. Default is unspecified (NULL).}

\item{verbose}{whether to print additional information about run (default: FALSE)}

\item{string.receptors}{a list of receptor names that STRING recognises (if cannot map defaults)}

\item{string.ligands}{a list of ligand names that STRING recegnises (if cannot map defaults)}

\item{string.input.map}{named vector of STRING-compatable gene symbols with names corresponding to genes in dataset}
}
\value{
a data-frame containing ligands, receptors and STRING association scores between them.
}
\description{
Generates a table of ligand:receptor association scores from the STRING database
}
\examples{

make_STRING_table(ligands, receptors, species="mouse")

## Or if STRING doesn't recognize some genes
string.input.map <- c("Ackr3)
names(string.input.map) <- c("Cxcr7")
string.receptors <- c("Cxcr7)
make_STRING_table(ligands, receptors, species="mouse", string.receptors=string.receptors, string.input.map=string.input.map)

}
