% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PlotTopLigands}
\alias{PlotTopLigands}
\title{Plot a bar graph of the top ligands}
\usage{
PlotTopLigands(
  input.file,
  cell.identity,
  col.use = "#757575",
  min.weight = 2,
  max.plot.num = 15,
  title.use = NULL
)
}
\arguments{
\item{input.file}{a table of source:ligand:receptor:target paths.}

\item{cell.identity}{the name of a cell population/cluster to use}

\item{col.use}{optional paramater to set colour of the bar graph.}

\item{min.weight}{minimum weight for including paths in the top ligand calculations.}

\item{max.plot.num}{the maximum number of ligands to plot}
}
\value{
a ggplot2 object.
}
\description{
Plot a bar graph of the top ligands
}
\examples{
\dontrun{
PlotTopLigands(path.table = example.table, cell.identity = '1')
}

}
